Check for process existence; be more verbose

+++ assessments/ch12/murder.c
@@ -63,10 +63,13 @@ int main(int argc, char **argv)
 
 	printf("Fair Warning given, proceeding to murder ...\n");
 	for (i=1; i<pid_max; i++) {
+		if (kill (i, 0) < 0)
+			continue;
 		if (kill (i, SIGKILL) < 0)
 			WARN("kill PID %d with SIGKILL failed\n", i);
+		printf(" SIGKILL -> %d\n", i);
 	}
-
+	
 	exit(EXIT_SUCCESS);
 }
 /* vi: ts=8 */
